Sets the "Copy Local" to false
In Visual Studio, by default, when you build a project, all the assembly
references are copied over to the target directory.

This is problematic for Umbraco packages, as if you XCOPY all the
Umbraco assemblies over to an existing Umbraco installation, then you
may end up with a version mismatch.
e.g. v8.0.0 assemblies in a v8.0.1 installation.
and Umbraco does not like that at all.

So we need to make sure that the Umbraco assemblies aren't in the
target directory.